SLING-10937 move test to different class to make sure it is executed for all resource resolver types (only RESOURCERESOLVER_MOCK fails)
